+++ libs/crypto/proof_of_work_test.go
@@ -5,10 +5,13 @@ import (
 	"testing"
 
 	"code.vegaprotocol.io/vega/libs/crypto"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestPoW(t *testing.T) {
+	t.Parallel()
+
 	_, _, err := crypto.PoW(crypto.RandomHash(), crypto.RandomHash(), 5, "nonExisting")
 	require.Error(t, err)
 
@@ -26,6 +29,8 @@ func TestPoW(t *testing.T) {
 }
 
 func TestVerify(t *testing.T) {
+	t.Parallel()
+
 	success, _ := crypto.Verify("", "", 0, "non existing", 0)
 	require.False(t, false, success)
 	success, _ = crypto.Verify("", "", 0, "non existing", 1)
@@ -47,24 +52,60 @@ func TestVerify(t *testing.T) {
 }
 
 func TestCountZeros(t *testing.T) {
-	// 00000e31f8ac983354f5885d46b7631bc75f69ec82e8f6178bae53db0ab7e054
-	_, h1, _ := crypto.PoW("2E7A16D9EF690F0D2BEED115FBA13BA2AAA16C8F971910AD88C72B9DB010C7D4", "DFE522E234D67E6AE3F017859F898E576B3928EA57310B765398615A0D3FDE2F", 20, crypto.Sha3)
-	require.Equal(t, byte(20), crypto.CountZeros(h1))
+	t.Parallel()
 
-	// 0000077b7d66117b57e45ccba0c31554e61c9853cc1cd9a2cf09c41b0aa9c22e
-	_, h2, _ := crypto.PoW("2E7A16D9EF690F0D2BEED115FBA13BA2AAA16C8F971910AD88C72B9DB010C7D4", "5B87F9DFA41DABE84A11CA78D9FE11DA8FC2AA926004CA66454A7AF0A206480D", 21, crypto.Sha3)
-	require.Equal(t, byte(21), crypto.CountZeros(h2))
+	tcs := []struct {
+		name       string
+		blockHash  string
+		txID       string
+		difficulty uint
+	}{
+		{
+			// 00000e31f8ac983354f5885d46b7631bc75f69ec82e8f6178bae53db0ab7e054
+			name:       "with difficulty set to 20",
+			blockHash:  "2E7A16D9EF690F0D2BEED115FBA13BA2AAA16C8F971910AD88C72B9DB010C7D4",
+			txID:       "DFE522E234D67E6AE3F017859F898E576B3928EA57310B765398615A0D3FDE2F",
+			difficulty: 20,
+		}, {
+			// 0000077b7d66117b57e45ccba0c31554e61c9853cc1cd9a2cf09c41b0aa9c22e
+			name:       "with difficulty set to 21",
+			blockHash:  "2E7A16D9EF690F0D2BEED115FBA13BA2AAA16C8F971910AD88C72B9DB010C7D4",
+			txID:       "5B87F9DFA41DABE84A11CA78D9FE11DA8FC2AA926004CA66454A7AF0A206480D",
+			difficulty: 21,
+		}, {
+			// 000003bbf0cde49e3899ad23282b18defbc12a65f07c95d768464b87024df368
+			name:       "with difficulty set to 22",
+			blockHash:  "2E7A16D9EF690F0D2BEED115FBA13BA2AAA16C8F971910AD88C72B9DB010C7D4",
+			txID:       "B14DD602ED48C9F7B5367105A4A97FFC9199EA0C9E1490B786534768DD1538EF",
+			difficulty: 22,
+		}, {
+			// 000001e1084b865aba27df7a445753a24a3d89d63c7739a62c11dab3ee6eae32
+			name:       "with difficulty set to 23",
+			blockHash:  "B14DD602ED48C9F7B5367105A4A97FFC9199EA0C9E1490B786534768DD1538EF",
+			txID:       "94A9CB1532011081B013CCD8E6AAA832CAB1CBA603F0C5A093B14C4961E5E7F0",
+			difficulty: 23,
+		},
+	}
 
-	// 000003bbf0cde49e3899ad23282b18defbc12a65f07c95d768464b87024df368
-	_, h3, _ := crypto.PoW("2E7A16D9EF690F0D2BEED115FBA13BA2AAA16C8F971910AD88C72B9DB010C7D4", "B14DD602ED48C9F7B5367105A4A97FFC9199EA0C9E1490B786534768DD1538EF", 22, crypto.Sha3)
-	require.Equal(t, byte(22), crypto.CountZeros(h3))
+	for _, tc := range tcs {
+		t.Run(tc.name, func(tt *testing.T) {
+			tt.Parallel()
 
-	// 000001e1084b865aba27df7a445753a24a3d89d63c7739a62c11dab3ee6eae32
-	_, h4, _ := crypto.PoW("B14DD602ED48C9F7B5367105A4A97FFC9199EA0C9E1490B786534768DD1538EF", "94A9CB1532011081B013CCD8E6AAA832CAB1CBA603F0C5A093B14C4961E5E7F0", 23, crypto.Sha3)
-	require.Equal(t, byte(23), crypto.CountZeros(h4))
+			_, hash, err := crypto.PoW(tc.blockHash, tc.txID, tc.difficulty, crypto.Sha3)
+
+			require.NoError(tt, err)
+			assert.NotEmpty(tt, hash)
+
+			zeros := crypto.CountZeros(hash)
+
+			require.Equal(tt, byte(tc.difficulty), zeros)
+		})
+	}
 }
 
 func TestDifficulty(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		name       string
 		difficulty uint
@@ -114,16 +155,18 @@ func TestDifficulty(t *testing.T) {
 			proof:      []byte("000007542dcb39d1471fd6c7424a547b9039382e055ceed10c839f2b76f88c0d"),
 		},
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			n, h, err := crypto.PoW(tt.blockHash, tt.tid, tt.difficulty, crypto.Sha3)
-			require.NoError(t, err)
-			require.Equal(t, tt.nonce, n)
-			require.Equal(t, string(tt.proof), hex.EncodeToString(h))
+	for _, tc := range tests {
+		t.Run(tc.name, func(tt *testing.T) {
+			tt.Parallel()
 
-			b, d := crypto.Verify(tt.blockHash, tt.tid, tt.nonce, crypto.Sha3, tt.difficulty)
-			require.Equal(t, true, b)
-			require.True(t, d >= byte(tt.difficulty))
+			n, h, err := crypto.PoW(tc.blockHash, tc.tid, tc.difficulty, crypto.Sha3)
+			require.NoError(tt, err)
+			require.Equal(tt, tc.nonce, n)
+			require.Equal(tt, string(tc.proof), hex.EncodeToString(h))
+
+			b, d := crypto.Verify(tc.blockHash, tc.tid, tc.nonce, crypto.Sha3, tc.difficulty)
+			require.Equal(tt, true, b)
+			require.True(tt, d >= byte(tc.difficulty))
 		})
 	}
 }